+++ pulldown-cmark-mdcat/src/theme.rs
@@ -31,6 +31,8 @@ pub struct Theme {
     pub(crate) heading_style: Style,
     /// Style for footnote references and definitions.
     pub(crate) footnote_style: Style,
+    /// Style for math expressions.
+    pub(crate) math_style: Style,
 }
 
 impl Default for Theme {
@@ -46,6 +48,7 @@ impl Default for Theme {
             code_block_border_color: AnsiColor::Green.into(),
             heading_style: Style::new().fg_color(Some(AnsiColor::Blue.into())).bold(),
             footnote_style: Style::new().fg_color(Some(AnsiColor::Cyan.into())),
+            math_style: Style::new().fg_color(Some(AnsiColor::Yellow.into())),
         }
     }
 }