+++ src/lib.rs
@@ -84,7 +84,8 @@ pub fn process_file(
         Options::ENABLE_TASKLISTS
             | Options::ENABLE_STRIKETHROUGH
             | Options::ENABLE_TABLES
-            | Options::ENABLE_FOOTNOTES,
+            | Options::ENABLE_FOOTNOTES
+            | Options::ENABLE_MATH,
     );
     let env = Environment::for_local_directory(&base_dir)?;